cambios de lugar de funcion delete
